Start doing #22
All I have to do now is make the changes communicate to the server
again, then I have to change the create project page to include the tag
component, and then I have to split the CSS, and then I have to fix all
the stuff I break. It's just that easy.